backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
langchain
langchain-community
pinecone-client==2.2.4
sentence-transformers
